backend/requirements.txt
Django==3.0.3
pytz==2019.3
sqlparse==0.3.0
djangorestframework==3.11.0
django-filter==2.2.0
Pillow==7.0.0
psycopg2-binary==2.8.4
gunicorn==20.0.4
django-cors-headers==3.2.1